Fixes for extended Limit command.

+++ trunk/src/sql.c
@@ -5591,7 +5591,7 @@ __C_LINK int __DPSCALL DpsSQLLimit8(DPS_AGENT *A, DPS_UINT8URLIDLIST *L, const c
 	u = 1;
 	offset = (dps_uint8)0;
 	while (u) {
-	  dps_snprintf(qbuf, qbuflen, "%s OFFSET %ld LIMIT %d", req, offset, url_num);
+	  dps_snprintf(qbuf, qbuflen, "%s OFFSET %ld LIMIT %d", req, (long)offset, url_num);
 	  for (i = 0; i < 3; i++) {
 	    if (A->flags & DPS_FLAG_UNOCON) DPS_GETLOCK(A, DPS_LOCK_DB);
 	    rc = DpsSQLQuery(db, &SQLres, qbuf);
@@ -5629,7 +5629,7 @@ __C_LINK int __DPSCALL DpsSQLLimit8(DPS_AGENT *A, DPS_UINT8URLIDLIST *L, const c
 	  }
 	  DpsSQLFree(&SQLres);
 	  offset += (dps_uint8)nrows;
-	  DpsLog(A, DPS_LOG_EXTRA, "%ld records processed.", offset);
+	  DpsLog(A, DPS_LOG_EXTRA, "%ld records processed.", (long)offset);
 	  L->nitems += p;
 	  u = (nrows == url_num);
 	}
@@ -5654,7 +5654,7 @@ __C_LINK int __DPSCALL DpsSQLLimit4(DPS_AGENT *A, DPS_UINT4URLIDLIST *L, const c
 	u = 1;
 	offset = (dps_uint8)0;
 	while (u) {
-	  dps_snprintf(qbuf, qbuflen, "%s OFFSET %ld LIMIT %d", req, offset, url_num);
+	  dps_snprintf(qbuf, qbuflen, "%s OFFSET %ld LIMIT %d", req, (long)offset, url_num);
 	  for (i = 0; i < 3; i++) {
 	    if (A->flags & DPS_FLAG_UNOCON) DPS_GETLOCK(A, DPS_LOCK_DB);
 	    rc = DpsSQLQuery(db, &SQLres, qbuf);
@@ -5705,7 +5705,7 @@ __C_LINK int __DPSCALL DpsSQLLimit4(DPS_AGENT *A, DPS_UINT4URLIDLIST *L, const c
 	  }
 	  DpsSQLFree(&SQLres);
 	  offset += (dps_uint8)nrows;
-	  DpsLog(A, DPS_LOG_EXTRA, "%ld records processed.", offset);
+	  DpsLog(A, DPS_LOG_EXTRA, "%ld records processed.", (long)offset);
 	  L->nitems += p;
 	  u = (nrows == url_num);
 	}
